requirement.txt
langchain
langchain_community
ipykernel
python-dotenv
pypdf
bs4
langchain-text-splitters
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_groq
langchain_core
fastapi
uvicorn
langserve
streamlit
chroma-hnswlib
chromadb
huggingface-hub
langchain-openai
langchainhub
langdetect
langsmith
numpy
openai
opencv-python
pandas
pdf2image
pdfminer.six
pdfplumber
pikepdf
pillow
pillow_heif
pyarrow
pydantic
pypdf2
pypdfium2
pytesseract
PyYAML
rank-bm25
regex
requests
rich
safetensors
scikit-learn
scipy
SQLAlchemy
sympy
tiktoken
tokenizers
toml
torch
torchvision
tqdm
transformers
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
watchdog
watchfiles
wikipedia
google-generativeai
langchain_google_genai
InstructorEmbedding
